Solving remark 1 of PR
