+++ server/app/router.py
@@ -3,11 +3,10 @@
 from fastapi import APIRouter
 
 from app import auth, submit, hello
-from app import vectors, backbones
+from app import vectors
 
 router = APIRouter()
 router.include_router(auth.router)
 router.include_router(submit.router)
 router.include_router(hello.router)
 router.include_router(vectors.router)
-router.include_router(backbones.router)
