+++ server/app/vectors.py
@@ -15,3 +15,11 @@ def get_vectors(
     current_user: schemas.User = Depends(deps.get_current_user),
 ) -> List[Vector]:
     return crud.get_level0_for_user(database=database, user=current_user)
+
+
+@router.get("/vectors/backbones", response_model=List[schemas.Vector])
+def get_backbones(
+    database: Session = Depends(deps.get_db),
+    current_user: schemas.User = Depends(deps.get_current_user),
+) -> List[Vector]:
+    return crud.get_backbones_for_user(database=database, user=current_user)
